Credentials code and docs cleanup
